Fixed is_Ack method but get_dependies metod not recursive. removed own tag because merged updated and maintained query in ack_nack

+++ staj-projeleri/noan/noan/repository/views.py
@@ -74,13 +74,16 @@ def page_binary(request, distName, distRelease, sourceName, packageName, binaryN
     package = Package.objects.get(name=packageName, source=source)
     binary = Binary.objects.get(no=binaryNo, package=package)
     if request.method == "POST" and request.user and request.user.is_authenticated():
-        if request.POST['result'] == "unknown":
-            TestResult.objects.filter(binary=binary, created_by=request.user).delete()
-        elif request.POST['result'] in ("yes", "no"):
-            result, created = TestResult.objects.get_or_create(binary=binary, created_by=request.user)
-            result.result = request.POST['result']
-            result.save()
-
+        copy = request.POST
+        if "result" in request.POST:
+            if request.POST['result'] == "unknown":
+                TestResult.objects.filter(binary=binary, created_by=request.user).delete()
+            elif request.POST['result'] in ("yes", "no"):
+                result, created = TestResult.objects.get_or_create(binary=binary, created_by=request.user)
+                result.result = request.POST['result']
+                result.save()
+        print request.POST
+    #"csrfmiddlewaretoken"
     user_result = "unknown"
     if request.user and request.user.is_authenticated():
         results = binary.testresult_set.filter(created_by=request.user)
@@ -182,7 +185,7 @@ def search_form(request):
 def AckNackList(request):
     list=[]
     error = ()
-    stateBinary = Binary.objects.filter(Q(resolution = 'pending'), Q(package__source__maintained_by = request.user) | Q(update__updated_by = request.user),Q(testresult__isnull=True))
+    stateBinary = Binary.objects.filter(Q(resolution = 'pending'), Q(package__source__maintained_by = request.user) | Q(update__updated_by = request.user), Q(testresult__isnull=True))
     if request.method == 'POST':
         radio = {}
         comment = {}
@@ -213,7 +216,6 @@ def AckNackList(request):
                 Add_Comment.save()
     if stateBinary:
         distributions = Distribution.objects.all()
-
     else:
         distributions = ""
     paginator = Paginator(stateBinary, 10)
